code optimization, closes #59
