+++ tg_bot_app.py
@@ -8,7 +8,4 @@ from app import create_app
 if __name__ == '__main__':
     app = create_app()
     with app.app_context():
-        from telebot import apihelper
-        apihelper.proxy = {"https": "socks5://127.0.0.1:9150"}
-        bot.remove_webhook()
         bot.polling(none_stop=True, interval=0)
